payment_api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
kafka-python==2.0.2
redis==4.5.5